Add support for defining 'lazy attributes' on nodes
These are predefined in the node attributes with a sentinel value.
The first the key is accessed (via the node), the attribute will be
evaluated. Each lazy attribute corresponds to a property or method on
the node instance, e.g. node['foo'] -> node.foo().

This ensures attributes are accessible on the node, but prevents costly
attributes to be evaluated on initialization.