Refactor of implementation to remove unnecessary abstraction
The origins.api.Node class has been removed and nodes are now exposed
as is. The abstraction originally was intended to provide a consistent
access layer for traversing nodes, however this has been achieved on the
origins.base.Node.

Conventions have been put in place to standardize the implementation of
backends. The design is still limited in the way nodes are populated with
attributes and how they expose branches and elements. There is quite a bit
of redundancy across nodes since all of the logic for establishing the
attributes have been moved to the client. However, the Origin node must
still synchronize with the client to populate it's attributes which is
confusing.