Introduce formal graph structures including Node and Rel
Previously the graph structure was inferred based on the structure
of the backend nodes such as Database -> Table -> Column. This is suitable
for purely hierarchical structures, however representing any other
relationship in this structure was not possible due to not having a formal
graph structure.

Further, elements and branch properties have been dropped in favor of
nodes having formal relationships which can be traversed generically.
Backends now define relationships and can utilize them by filtering
relationships based on their properties.