Refactor export API to be classed based for incremental preparation
The export API now requires targets to define an Exporter class that
requires two methods, `prepare` and `export`, to be implemented. This
approach enables incrementally preparing the export to a desired state
then exporting it to the actual target. `export` can be called many times
for different targets (separate servers or output files).