Introduce sections into redcap hierarchy
Modify the Form node class in _redcap so that form.fields will return
all fields in all sections of the form and form.sections will return all
sections. Create the Section node class in _redcap so that
section.fields will return all fields in the section. Modify the sync
method on Form to call client.sections and create the sync method on
Section to call client.fields, passing the parent form name in addition
to the section name.

Modify all three backend Client classes by adding a sections method that
returns all unique sections found on a form, and always returns at least
a section named 'default' (since sections are optional in redcap).
Modify all three backend Client classes' fields method to take a section
name in addition to the form name and filter the results to only include
fields in that section (when 'default' is passed as the section name,
all fields before the first defined section are returned).