Refactor provenance graph loader
This removes the Writer classes in favor of a function that prepares
the statements to be loaded.

In addition, support for referencing provenance descriptions can be done
by pre-defining the `origins:neo4j` or `origins:uuid` attribute. This
enables creating provenance documents that augment the existing graph
rather than creating new nodes each time and having to reconcile them
later.

The provenance models have been removed in favor of a simple set of maps
and sets.

The `origins:Dependence` relation has been integrated which provides a
model for representing dependence between entities.

Fix #43

Signed-off-by: Byron Ruth <b@devel.io>