Implement ResourceExporter and convenience methods
This includes a general class, convenience function and method on
the Node class for exporting backend data as resources for the Origins
graph.

The base class can be subclassed to alter the way component and
relationship data are exported and identified. The class supports
incremental exporting as well as the ability to dictate the direction
and type of relationships to export during traversal.

Fix #38
Fix #54

Signed-off-by: Byron Ruth <b@devel.io>