Update edge operations to create native edge
The native edge between the start and end nodes is now created when an
edge is added. Likewise the native edge is deleted when the edge is
removed. Therefore the native edges represents current state of
the graph.

This feature is added to support the native structure of the graph.
This combined with the ability to set node labels and the property
packing make it possible to execute raw queries as one would expect.

Signed-off-by: Byron Ruth <b@devel.io>