Add function for basic search
A predicate dict is passed with attribute names to be searched and
the value to be matched. String values will be matched using a regexp
operator. Non-string values will test for equality. The default operator
between expressions is ‘AND’, but can also be ‘OR’ or ‘XOR’ all with an
optional ‘NOT’ for negation.

Grouped expressions are not currently supported.

See: http://docs.neo4j.org/chunked/stable/query-where.html for details.

Signed-off-by: Byron Ruth <b@devel.io>