Implement storage package
storage.Store provides a simple interface for reading and writing facts
to storage. It also acts as a in-process cache for the data.

The store interacts with a implementation of storage.Engine for reading
and writing key/value data.

The storage.Codec interface defines the Marshal and Unmarshal methods
that encodes and decodes a fact into bytes, respectively. This is used
by the store so the underlying engine only has to work with bytes.

The default codec is using protocol buffers. The other codec are
there for reference at the moment.

The Engine.SetMany method takes advantage of storage engines that
support transactions or atomic multi-key operations such as boltdb and
sqlite.

A storage.Options type has been added to standardize passing in options
to storage engines. The engines will used the options it requires and
ignore the rest. This makes it easier to interface with.

Signed-off-by: Byron Ruth <b@devel.io>