Implement storage.Reader
This defines an explicit start and stop index for the partition so
readers read at any point in time after it was initialized will also
contain the same slice of facts. Otherwise new segments could be written
to the partition which would be leaked into existing redaers.

The Reader type defines (Min|Max)Time and (Start|End)Time methods for
accessing various time points.

Store.RangeReader takes two int64 timestamps used to define the range of
the partition to read.

Signed-off-by: Byron Ruth <b@devel.io>