Implement fact package
The Fact is the EAV-based data model that includes time to make it
semantically immutable. Facts themselves are stored in a particular
domain which provides isolation.

The fact.Reader interface defines a Read method that emulates io.Reader.
It reads data from the underlying stream, converts the data into to Fact
values, and fills the passed Facts slice with produced facts.

The "cvs" and "jsonstream" modules contain implementations of fact.Reader.
"csv" is a table-based format, while "jsonstream" is a newline-delimited
stream of JSON encoded facts.

Signed-off-by: Byron Ruth <b@devel.io>