Add support for a config file
This is purely for convenience and removes the need to specify the
required global options for the storage engine.

By default, if the file named `origins.json` or `origins.yml` exists in
the current working directory, it will be loaded automatically.
Otherwise the `--config` argument can be used to specify an alternate
location of the file.

Signed-off-by: Byron Ruth <b@devel.io>