Add working prototype of duplicate fact prevention
This implementation is not performant and only performs an exact match
check. It does not take in account if the EAV values are the same, but
the valid time is different. Furthermore, it does not handle the case if
an EAV triple is asserted again after being retracted in a previous
transaction.

Signed-off-by: Byron Ruth <b@devel.io>