Implement engine.Tx interface for storage engine
A transaction enables performing get, set, and incr atomically. The
actual guarantees are dependent on the engine implementation.

Engine.SetMany was replaced with Engine.Multi which takes a function
that has a transaction passed into it. This enables performing multiple
commands in the context of a transaction.

A general purpose engine test suite has been added to reduce redundancy
and ensure all implementations are tested equally.

Signed-off-by: Byron Ruth <b@devel.io>