Implement fact sort and comparators functions
This includes simple comparators for entity, attribute, value,
and time as well as multi-dimension comparators. The multi-dimension
comparators can be used to sort facts in various ways for building
various indexes for efficient access.

The EAVT index is used to efficiently access facts about a particular
entity since they are sorted by the entity id.

The AEVT index provides columnar-style access to the day, i.e. find
all the entities with some attribute.

The AVET index is used to find entities with attribute-value combinations.

The VAET index is used to find entities given some entity-based value.
This is used to perform reverse lookups of a relationship between two
entities.

Signed-off-by: Byron Ruth <b@devel.io>