Implement the key-value based storage layer
A key-value based storage layer is used for flexibility and efficiency
when reading and writing facts. In addition, existing key-value stores
can be utilized many of which are distributed systems to scale
reads and writes transparently.

The interface defines very basic methods including `Get`, `Set`,
`Delete`, and `Incr`. `Multi` takes a function that receives a
Tx value for encapsulating multiple operations. For implementations
that support transaction this provides a means of exposing that
functionality.

Signed-off-by: Byron Ruth <b@devel.io>