Implement transactor package
The transactor package implements transaction processing for facts. The
usage is straightforward; initialize a transaction, append facts to the
transaction, and commit it. On commit, the root of each affected
domain’s commit log will be updated to point to the transacted segment.

The time functions have been moved to the chrono package for shared
access by the origins and transactor packages.

Signed-off-by: Byron Ruth <b@devel.io>