Add option to toggle setting default fact time
The fact time, more generally known as the "valid time", defines when
the fact was valid in the real world. This change removes the assumption
that if a valid time is not set the fact is valid at the start of the
transaction. An unset valid time could mean it is unknown when the fact
became true in the real world. For this reason, a transaction option has
been added to toggle the behavior. The default behavior is to *not* set
the valid time to the transaction time.

Signed-off-by: Byron Ruth <b@devel.io>