Rename origins.Open to origins.Init
The return value is a storage engine and initialize better conveys
the operation taking place. Whether or not something is being opened
(file descriptor, network connection) is an implementation detail of
the engine itself. Initializing a storage engine typically creates any
base files or ensures a connection can be made.

Signed-off-by: Byron Ruth <b@devel.io>