Refactor block encoding/decoding
This change introduces a BlockEncoder and BlockDecoder which
incrementally encodes and decodes facts. The previous behavior was to
handle the whole block of facts at once which is more of a “stop-the-world”
approach.

Although the overall performance may not improved, the flow is much more
clear and incremental. In addition, this stream based composition is
necessary for refactoring the processing into a concurrent model.

Additional refactoring has been made to standardize on the Writer and
Flusher interfaces rather than using alternate methods (Append).

Increase maximum fact size has been increased to 2^32 bytes which
requires a 4-byte encoding prefix. Fix #146

Signed-off-by: Byron Ruth <b@devel.io>