Add iterator-based utility functions
Included are functions that return the unique entity, attribute,
or value identities and Filter which is a function that takes a fact
filter function. Also included is a Slice function that mimics applying
an offset and limit constraint on an iterator.

A new Idents type has been added that satisfies the sort.Sorter
interface. In addition, the comparator functions are now exposed for use
by other packages.

Signed-off-by: Byron Ruth <b@devel.io>