Add iterator utilities
The Map function takes an iterator and function that processes the fact
and returns and error. This provides a concise way of consuming an
iterator.

Entity takes an iterator and ident and yields only facts pertaining
to the entity.

Exists takes an iterator and a filter function and returns true if a
fact is returned that matches.

Signed-off-by: Byron Ruth <b@devel.io>