Add logic in transactor pipeline to prevent writing redundant facts
When a pipeline receives a fact, it will lookup the fact to determine
if it is redundant. This is done by building a local index of existing
entities in the domain as of the time of this transaction and evaluating
the current fact against the set.

Building the local index has an O(n) complexity with respective to
domain size for both time and space. This is required because the
underlying data must be loaded, sorted, and grouped to build the index.
However, once built, the lookup is an O(1) operation.

Future optimizations may write the index to temporary storage to reduce
in-process memory for long-running or multi-domain transactions.
Alternately, defining *hints* that the transaction caller could provide
about the properties of the stream could make the memory management
smarter.

The index is a trie for indexing keys that are strings and depends on
an external library that does not support Go 1.2 or below. This has been
removed from the CI configuration.

Signed-off-by: Byron Ruth <b@devel.io>