Bind to separate viper keys for storage options
If the same flag is associated with several cobra commands, the flag
should be bound to a different viper tag for each command.
Otherwise only the command that got initialized last gets to keep the
binding, and the other commands get stuck with the default value for
the flag.

In our case, init() gets called last for the transact command, so
transact was picking up the correct value of the --storage flag, and
other commands, like http and domains, were getting the default
storage==“memory” no matter what value was being passed in.