Refactor Reflect internals and add Identifier interface
The `Identifier` interface enables structs to specify the `Ident` they
are identified by. For example a typical `User` type in an application
may be uniquely defined by it’s `Username` or `Email` address field. The
`Ident()` can be implemented that returns an `Ident` with the value of
the identifying field.

Signed-off-by: Byron Ruth <b@devel.io>