Bind command flags in run phase
This fixes the problem of each subcommand overriding the flag that is
bound to a shared viper setting without needing to namespace them. Adding
a namespace prevented using a common name in configuration files or
environment variables.

Binding to the target flag is now deferred until the command’s Run
method is called which ensures it read from the correct flag set.

Fix #167

Signed-off-by: Byron Ruth <b@devel.io>